Moving rule to our encore pack. Changing file names, changing scripts to use the same error retrieval methods that now exist in actions/lib/base_action
