+++ actions/lib/base_action.py
@@ -14,6 +14,8 @@
 # limitations under the License.
 
 import socket
+import re
+from six import string_types
 from st2common.runners.base_action import Action
 from st2client.client import Client
 
@@ -26,6 +28,9 @@ class BaseAction(Action):
         :returns: a new BaseAction
         """
         super(BaseAction, self).__init__(config)
+        self.child_error = []
+        self.parent_output = []
+        self.errors_as_string = ""
 
     def st2_client_initialize(self, st2_exe_id):
         st2_fqdn = socket.getfqdn()
@@ -35,6 +40,140 @@ class BaseAction(Action):
         vm_execution = self.st2_client.executions.get_by_id(st2_exe_id)
 
         return vm_execution
+
+    def find_error_execution(self, parent_execution, ignored_error_tasks):
+        if hasattr(parent_execution, 'children'):
+            for m in parent_execution.children:
+                self.find_error_execution(m, ignored_error_tasks)
+        else:
+            st2_executions = self.st2_client.executions
+            execution = st2_executions.get_by_id(parent_execution)
+            if (execution.status == "failed" and
+               execution.context['orquesta']['task_name'] not in ignored_error_tasks):
+                execution_result = execution.result
+                if self.check_custom_errors(execution_result, execution):
+                    return None
+                self.parent_error = execution
+                if hasattr(execution, 'children'):
+                    for c in execution.children:
+                        self.find_error_execution(c, ignored_error_tasks)
+                else:
+                    self.child_error.append(execution)
+
+    def check_custom_errors(self, execution_result, execution):
+        if 'output' in execution_result and execution_result['output']:
+            if execution_result.get('output', {}).get('error'):
+                self.parent_error = execution
+                parent_errors = execution_result['output']['error']
+                if isinstance(parent_errors, string_types):
+                    self.errors_as_string = parent_errors
+                    return True
+                for error in parent_errors:
+                    for errors in error:
+                        self.parent_output.append(error['error'])
+                self.errors_as_string = '\n'.join(self.parent_output)
+                return True
+        return False
+
+    def format_error(self, html_tags):
+        err_string = ""
+
+        if self.child_error:
+            for error in self.child_error:
+                if html_tags:
+                    err_message = self.format_error_strings(self.get_error_message(error.result))
+                else:
+                    err_message = self.get_error_message(error.result)
+
+                err_string += self.get_error_string(html_tags,
+                                                    error.context['orquesta']['task_name'],
+                                                    error.id,
+                                                    err_message)
+        else:
+            if html_tags:
+                error_result = self.parent_error.result
+                if self.errors_as_string:
+                    parent_error = self.errors_as_string
+                else:
+                    parent_error = self.get_error_message(error_result)
+                err_message = self.format_error_strings(parent_error)
+                err_string += self.get_error_string(html_tags,
+                                                    self.parent_error
+                                                    .context['orquesta']['task_name'],
+                                                    self.parent_error.id,
+                                                    err_message)
+            else:
+                if self.errors_as_string:
+                    parent_error = self.errors_as_string
+                else:
+                    parent_error = self.get_error_message(self.parent_error.result)
+
+                err_string += self.get_error_string(html_tags,
+                                                    self.parent_error
+                                                    .context['orquesta']['task_name'],
+                                                    self.parent_error.id,
+                                                    parent_error)
+
+        return err_string
+
+    def get_error_string(self, html_tags, err_context, err_id, err_message):
+        if html_tags:
+            return ("Error task: {0}<br>Error execution ID: {1}<br>Error message: {2}"
+                    "<br>".format(err_context, err_id, err_message))
+        else:
+            return ("Error task: {0}\nError execution ID: {1}\nError message: {2}"
+                    "\n".format(err_context, err_id, err_message))
+
+    def format_error_strings(self, error_string):
+        """ formats error strings by dropping extra string escapes
+        and drops the ansi escapes. Then we convert the new lines into
+        <br> so that new lines appear correctly.
+        """
+        # ansi escape sequence
+        ansi_escape = re.compile(r'\x1B\[[0-?]*[ -/]*[@-~]')
+
+        while "\\n" in error_string:
+            # We need to encode first for python3 to decode byte string
+            error_string_encoded = error_string.encode()
+            error_string = error_string_encoded.decode('unicode_escape')
+
+        error_string = ansi_escape.sub('', error_string)
+
+        error_string = error_string.replace('\n', '<br>')
+
+        return error_string
+
+    def get_error_message(self, error_result):
+        # Custom Error Messages returned from workflow outputs
+        if 'output' in error_result and error_result['output']:
+            if 'error' in error_result['output']:
+                return error_result['output']['error']
+
+        # Jinja syntax errors
+        if 'errors' in error_result:
+            return error_result['errors'][0]['message']
+
+        # Bolt plans (https://github.com/StackStorm-Exchange/stackstorm-bolt)
+        if ('result' in error_result and
+           error_result['result'] and
+           error_result['result'] != 'None'):
+            if 'details' in error_result['result']:
+                if 'result_set' in error_result['result']['details']:
+                    result_set = error_result['result']['details']['result_set'][0]
+                    return result_set['value']['_error']['msg']
+
+            if isinstance(error_result['result'], string_types):
+                return error_result['result']
+
+            if 'stderr' in error_result['result']:
+                return error_result['result']['stderr']
+
+        # python actions
+        # ex. (vsphere pack https://github.com/StackStorm-Exchange/stackstorm-vsphere)
+        if 'stderr' in error_result:
+            return error_result['stderr']
+
+        return "Could not retrieve error message"
 
     def run(self, **kwargs):
         raise RuntimeError("run() not implemented")
