--- actions/provision_errors.yaml
@@ -1,12 +1,0 @@
----
-description: "Gives alerts when action failures occur during provisioning"
-enabled: true
-runner_type: "python-script"
-entry_point: provision_errors.py
-name: provision_errors
-pack: errors
-parameters:
-  st2_exe_id:
-    type: string
-    description: "Stackstorm execution id"
-    required: true
