--- rules/servicenow_event.yaml
@@ -1,18 +1,0 @@
----
-  name: "servicenow_event"
-  pack: "errors"
-  description: "Creates a new event in ServiceNow"
-  enabled: true
-
-  trigger:
-    type: "errors.update_events"
-    parameters: {}
-
-  action:
-    ref: "servicenow_encore.add_event"
-    parameters:
-      st2_comments: "{{ trigger.st2_comments }}"
-      st2_execution_id: "{{ trigger.st2_execution_id }}"
-      st2_rule_name: "{{ trigger.st2_rule_name }}"
-      st2_server: "{{ trigger.st2_server }}"
-      st2_state: "{{ trigger.st2_state }}"
